Improve print order and verbosity
Try to order by the RHS, since we assume that the user is looking at
that as the "new" version.  Insert "removed" (LHS-only) commits once
we have shown all the other commits that came before them on the LHS.
That way the LHS is in order "as far as it can be done".