Weigh creation patches by a constant factor
It turns out that giving creations (i.e. commits missing on one side)
full weight is actually not brilliant; they become so expensive that
it is better to match up two almost-unrelated commits.  So make it
configurable, and tentatively put the weight at 0.6 so that the sum of
two creations is somewhat above half a rewrite.