Test numpy array length explicitly
As Johan pointed out, we would sometimes not actually show the removed
commits.  This happened because if the very first commit has been
removed (and everything else depends on it), we would end up computing
the list of commits-to-print as [0] (i.e. only the index 0).

But testing that by implicit bool() doesn't do the right thing.  It's
a numpy array, and numpy is dumb, to wit:

  >>> bool([0])
  True
  >>> bool(np.array([0]))
  False

So ask for the len() explicitly.