Format an absent commit as _seven_ dashes
We format the hash by asking git for an abbreviated sha1 and stuffing
it into a %8s format.  So usually (depending on configs of course) it
will be 7 chars wide, with some space for a slightly longer hash.

This changes the "absent commit" dashes to also be seven wide, so that
they line up visually in the common case (in git.git anyway) that all
the commits are unique at 7 digits.

Of course we should eventually fix it to correctly prepare the output
in one go, formatting to the maximum width of all the hashes.