Ignore separator line between commits
We didn't have any code for the separating blank line between commits
in git-log output, so it showed up as a trailing blank line in the
diffs.  This would show a spurious change if the last commit was
reordered to another (not last) position, because the last commit does
not have this blank line.

Skip it during diff reading.  This seems a reasonable place since it
would also catch other separating blank lines within the diff (not
that we should ever get those), and fixes the problem.